backend/requirements.txt
# Django Core
Django==5.2.1
djangorestframework==3.15.2
django-cors-headers==4.6.0
django-environ==0.11.2

# Authentication
djangorestframework-simplejwt==5.3.1
django-ratelimit==4.1.0
django-ratelimit-backend==0.2.0

# Database
psycopg2-binary==2.9.9
pymongo==4.10.1
django-filter==24.3

# Task Queue
celery==5.4.0
redis==5.2.0

# ML/NLP
scikit-learn==1.5.2
tensorflow==2.18.0
transformers==4.46.3
sentence-transformers==3.3.1
spacy==3.8.2
numpy==2.1.3
pandas==2.2.3
joblib==1.4.2

# Web Scraping
beautifulsoup4==4.12.3
selenium==4.27.1
requests==2.32.3
lxml==5.3.0

# Utilities
python-dateutil==2.9.0
pytz==2024.2

# Testing
pytest==8.3.4
pytest-django==4.9.0
pytest-cov==6.0.0
coverage==7.6.1

# Production
gunicorn==23.0.0
whitenoise==6.8.2

